do not fork into background with any fat binary
this avoids that someone e.g. porting to freebsd (which is also
supported by PyInstaller) will stumble over the platform check.

likely this is a very general problem with the fat binaries:

i suspect it is due to the pyinstaller bootloader / wrapper
cleaning up the temporary directory _MEIxxxxx where it extracted
all the needed files after the main process terminates (sys.exit()),
being unaware that we have just generated a background process also
relying on these files.

for the linux case, this led to some strange exception from pytz
that the current timezone is not supported (likely because it failed
to load some pytz tz definition file which was already gone).