Improvements to the mount/unmount feature
Now the mount point of an archive is stored per archive instead of just
one global mount point. This allows to mount more than one archive at
the same time.

When the user selects an archive, the mount button is updated and
changes between Mount/Unmount actions depending on the state of the
selected archive.

Also, now the archive name is stored in the action's params dictionary
since it's needed in the function processing a successful mount/unmount.