Add a context menu to the archive table (and open folder action)
This adds a context menu to the archives in the archive table which by
default just provides the same actions available as buttons below, but
when an archive is mounted, also provides an "Open Folder..." action.
A way to quickly open a file browser on the archive's mount point.

The open folder action was tested on Linux, where it opens the file
browser defined for the desktop (dolphin on KDE, nautilus on GNOME, for
example). I also tried to make it work on macOS, but some testing would
be nice since I don't have such a machine.