Show/Hide the main window when clicking on the tray icon
This makes it easier to open the Vorta main window just by clicking
on the tray icon (instead of having to right-click and then click on
Settings). Also, vorta now has the common behaviour of closing the
main window if it's already visible when the user clicks on the tray
icon.