Prevent duplicate repos from being added. Fixes #473
The change to repo_add_dialog.py adds a check in the database
to see if the repo already exists, and blocks the ui with an
error message if so.
The change to repo_tab.py refactors the selector population logic
so that the selector stays in sync with the database and doesn't
leave the entry in the UI that triggers the crash.