Add a development mode (#1162)
This creates the option to use the --development or -D flags
to launch vorta in development mode which stores all settings,
cache, temp, and log files in the project tree instead of in the
default places for the system.

Allows for working on the project without it pulling in the settings
from the system-wide copy of vorta that you use for your actual backups.