Fix issues as discussed in #1682.
Changes based on the discussion here:
https://github.com/borgbase/vorta/pull/1682

Biggest change is that src/config is rewritten to initialize
the config variables while the program is running, so you need
to import the whole module every time you use variables/constants
from it.

This was done to avoid duplicating code responsible for setting
up the directory in which config data will be stored.